+++ modules/base/GSConst.mjs
@@ -1,0 +1,13 @@
+/*
+ * Copyright (C) 2015, 2026; Green Screens Ltd.
+ */
+
+export const KEY = Symbol.for('gs-element');
+export const OWNER = Symbol.for('gs-owner');
+export const PARENT = Symbol.for('gs-parent');
+export const HANDLER = Symbol.for('gs-handler');
+export const EVENTID = Symbol.for('gs-event-id')
+export const NEXT = Symbol.for('gs-next');
+export const PREV = Symbol.for('gs-prev');
+export const RENDER = Symbol.for("GS-TEMPLATE-RENDER");
+export const DEF = Symbol.for('#def');