+++ modules/components/Dialog.mjs
@@ -3,7 +3,8 @@
  */
 
 import { classMap, createRef, css, html, ifDefined, ref } from '../lib.mjs';
-import { GSElement, HANDLER_KEY } from '../GSElement.mjs';
+import { HANDLER } from '../base/GSConst.mjs';
+import { GSElement } from '../GSElement.mjs';
 import { GSEvents } from '../base/GSEvents.mjs';
 import { GSDOM } from '../base/GSDOM.mjs';
 import { GSUtil } from '../base/GSUtil.mjs';
@@ -159,20 +160,20 @@ export class GSDialogElement extends GSElement {
 
   addController(controller) {
     if (controller.isDialog) {
-      this[HANDLER_KEY]?.add(controller);
+      this[HANDLER]?.add(controller);
     }
     super.addController?.(controller);
   }
 
   afterOpen() {
     const me = this;
-    me[HANDLER_KEY]?.forEach((c) => c.afterOpen?.(me));
+    me[HANDLER]?.forEach((c) => c.afterOpen?.(me));
     return me.emit('open', null, false, false, true);
   }
 
   afterClose() {
     const me = this;
-    me[HANDLER_KEY]?.forEach((c) => c.afterClose?.(me));
+    me[HANDLER]?.forEach((c) => c.afterClose?.(me));
     return me.emit('close', null, false, false, true);
   }