+++ modules/components/ext/GSTextAreaElement.mjs
@@ -6,7 +6,7 @@
  * A module loading GSDataListExt class
  * @module components/ext/GSDataListExt
  */
-
+import { OWNER, PARENT } from "../../base/GSConst.mjs";
 import { GSDOM } from "../../base/GSDOM.mjs";
 import { GSLoader } from "../../base/GSLoader.mjs";
 import { GSEvents } from "../../base/GSEvents.mjs";
@@ -116,11 +116,11 @@ export class GSExtTextAreaElement extends HTMLTextAreaElement {
     }
 
     get owner() {
-        return this[Symbol.for('gs-owner')]();
+        return this[OWNER]();
     }
 
     get parentComponent() {
-        return this[Symbol.for('gs-parent')]();
+        return this[PARENT]();
     }
 
     get hasUpdated() {